// src/app/models/imovel.ts
export class Imovel{
    id: number;
    nomeContribuinte: string;
    cpfContribuinte: string;
    logradouroImovel: string;
    numeroImovel: string;
    complementoImovel: string;
    cepImovel: string;
    bairroImovel: string;
    setor: string;
    quadra: string;
    lote: string;
    logradouroContribuinte: string;
    numeroContribuinte: string;
    complementoContribuinte: string;
    bairroContribuinte: string;
    cepContribuinte: string;
    cidadeContribuinte: string;
    valorVenalEdificacao: string;
    valorVenalTerreno: string;
    valorVenalImovel: string;
    totalAreaEdificada: string;
    areaTerreno: string;

}